+++ e2e/detect-open-handles/__tests__/promise.js
@@ -1,3 +1,5 @@
+// Copyright (c) 2014-present, Facebook, Inc. All rights reserved.
+
 test('something', () => {
   // eslint-disable-next-line no-new
   new Promise(() => {});