+++ packages/jest-config/src/ValidConfig.js
@@ -10,7 +10,7 @@
 import type {InitialOptions} from 'types/Config';
 
 import {replacePathSepForRegex} from 'jest-regex-util';
-import {MultipleValidOptions} from 'jest-validate';
+import {multipleValidOptions} from 'jest-validate';
 import {NODE_MODULES} from './constants';
 
 const NODE_MODULES_REGEXP = replacePathSepForRegex(NODE_MODULES);
@@ -40,6 +40,7 @@ export default ({
       statements: 100,
     },
   },
+  dependencyExtractor: '<rootDir>/dependencyExtractor.js',
   displayName: 'project-name',
   errorOnDeprecated: false,
   expand: false,
@@ -50,6 +51,7 @@ export default ({
   globalTeardown: 'teardown.js',
   globals: {__DEV__: true},
   haste: {
+    computeSha1: true,
     defaultPlatform: 'ios',
     hasteImplModulePath: '<rootDir>/haste_impl.js',
     platforms: ['ios', 'android'],
@@ -101,7 +103,7 @@ export default ({
   testMatch: ['**/__tests__/**/*.js?(x)', '**/?(*.)+(spec|test).js?(x)'],
   testNamePattern: 'test signature',
   testPathIgnorePatterns: [NODE_MODULES_REGEXP],
-  testRegex: MultipleValidOptions(
+  testRegex: multipleValidOptions(
     '(/__tests__/.*|(\\.|/)(test|spec))\\.jsx?$',
     ['/__tests__/\\.test\\.jsx?$', '/__tests__/\\.spec\\.jsx?$'],
   ),