+++ packages/jest-config/src/__tests__/normalize.test.js
@@ -671,11 +671,8 @@ describe('testEnvironment', () => {
   beforeEach(() => {
     Resolver = require('jest-resolve');
     Resolver.findNodeModule = jest.fn(name => {
-      if (name === 'jsdom') {
-        return 'node_modules/jsdom';
-      }
-      if (name === 'jest-environment-jsdom') {
-        return 'node_modules/jest-environment-jsdom';
+      if (['jsdom', 'jest-environment-jsdom'].includes(name)) {
+        return `node_modules/${name}`;
       }
       if (name.startsWith('/root')) {
         return name;
@@ -860,7 +857,7 @@ describe('testRegex', () => {
 
     expect(options.testRegex).toEqual([]);
   });
-  it('testRegex string is mapped to array of RegExp objects', () => {
+  it('testRegex string is mapped to an array', () => {
     const {options} = normalize(
       {
         rootDir: '/root',
@@ -869,9 +866,9 @@ describe('testRegex', () => {
       {},
     );
 
-    expect(options.testRegex).toEqual([/.*/]);
+    expect(options.testRegex).toEqual(['.*']);
   });
-  it('testRegex array is mapped to array of RegExp objects', () => {
+  it('testRegex array is preserved', () => {
     const {options} = normalize(
       {
         rootDir: '/root',
@@ -880,7 +877,7 @@ describe('testRegex', () => {
       {},
     );
 
-    expect(options.testRegex).toEqual([/.*/, /foo\.bar/]);
+    expect(options.testRegex).toEqual(['.*', 'foo\\.bar']);
   });
 });
 
@@ -916,18 +913,6 @@ describe('testMatch', () => {
           rootDir: '/root',
           testMatch: ['**/*.js'],
           testRegex: '.*',
-        },
-        {},
-      );
-    }).toThrowErrorMatchingSnapshot();
-  });
-
-  it('throws if testRegex is provided an invalid regex string', () => {
-    expect(() => {
-      normalize(
-        {
-          rootDir: '/root',
-          testRegex: 'foo(bar',
         },
         {},
       );
@@ -1217,15 +1202,79 @@ describe('preset without setupFiles', () => {
   });
 });
 
+describe('runner', () => {
+  let Resolver;
+  beforeEach(() => {
+    Resolver = require('jest-resolve');
+    Resolver.findNodeModule = jest.fn(name => {
+      if (['eslint', 'jest-runner-eslint', 'my-runner-foo'].includes(name)) {
+        return `node_modules/${name}`;
+      }
+      if (name.startsWith('/root')) {
+        return name;
+      }
+      return findNodeModule(name);
+    });
+  });
+
+  it('defaults to `jest-runner`', () => {
+    const {options} = normalize({rootDir: '/root'}, {});
+
+    expect(options.runner).toBe('jest-runner');
+  });
+
+  it('resolves to runners that do not have the prefix', () => {
+    const {options} = normalize(
+      {
+        rootDir: '/root/',
+        runner: 'my-runner-foo',
+      },
+      {},
+    );
+
+    expect(options.runner).toBe('node_modules/my-runner-foo');
+  });
+
+  it('resolves to runners and prefers jest-runner-`name`', () => {
+    const {options} = normalize(
+      {
+        rootDir: '/root/',
+        runner: 'eslint',
+      },
+      {},
+    );
+
+    expect(options.runner).toBe('node_modules/jest-runner-eslint');
+  });
+
+  it('throw error when a runner is not found', () => {
+    expect(() =>
+      normalize(
+        {
+          rootDir: '/root/',
+          runner: 'missing-runner',
+        },
+        {},
+      ),
+    ).toThrowErrorMatchingSnapshot();
+  });
+});
+
 describe('watchPlugins', () => {
   let Resolver;
   beforeEach(() => {
     Resolver = require('jest-resolve');
     Resolver.findNodeModule = jest.fn(name => {
-      if (name.startsWith(path.sep)) {
+      if (
+        ['typeahead', 'jest-watch-typeahead', 'my-watch-plugin'].includes(name)
+      ) {
+        return `node_modules/${name}`;
+      }
+
+      if (name.startsWith('/root')) {
         return name;
       }
-      return path.sep + 'node_modules' + path.sep + name;
+      return findNodeModule(name);
     });
   });
 
@@ -1235,19 +1284,59 @@ describe('watchPlugins', () => {
     expect(options.watchPlugins).toEqual(undefined);
   });
 
-  it('normalizes watchPlugins', () => {
+  it('resolves to watch plugins and prefers jest-watch-`name`', () => {
     const {options} = normalize(
       {
         rootDir: '/root/',
-        watchPlugins: ['my-watch-plugin', '<rootDir>/path/to/plugin'],
+        watchPlugins: ['typeahead'],
       },
       {},
     );
 
     expect(options.watchPlugins).toEqual([
-      {config: {}, path: '/node_modules/my-watch-plugin'},
+      {config: {}, path: 'node_modules/jest-watch-typeahead'},
+    ]);
+  });
+
+  it('resolves watch plugins that do not have the prefix', () => {
+    const {options} = normalize(
+      {
+        rootDir: '/root/',
+        watchPlugins: ['my-watch-plugin'],
+      },
+      {},
+    );
+
+    expect(options.watchPlugins).toEqual([
+      {config: {}, path: 'node_modules/my-watch-plugin'},
+    ]);
+  });
+
+  it('normalizes multiple watchPlugins', () => {
+    const {options} = normalize(
+      {
+        rootDir: '/root/',
+        watchPlugins: ['jest-watch-typeahead', '<rootDir>/path/to/plugin'],
+      },
+      {},
+    );
+
+    expect(options.watchPlugins).toEqual([
+      {config: {}, path: 'node_modules/jest-watch-typeahead'},
       {config: {}, path: '/root/path/to/plugin'},
     ]);
+  });
+
+  it('throw error when a watch plugin is not found', () => {
+    expect(() =>
+      normalize(
+        {
+          rootDir: '/root/',
+          watchPlugins: ['missing-plugin'],
+        },
+        {},
+      ),
+    ).toThrowErrorMatchingSnapshot();
   });
 });