+++ packages/jest-validate/src/__tests__/validate.test.js
@@ -9,15 +9,15 @@
 'use strict';
 
 import validate from '../validate';
-import {MultipleValidOptions} from '../condition';
-import jestValidateExampleConfig from '../example_config';
-import jestValidateDefaultConfig from '../default_config';
+import {multipleValidOptions} from '../condition';
+import jestValidateExampleConfig from '../exampleConfig';
+import jestValidateDefaultConfig from '../defaultConfig';
 
 const {
   defaultConfig,
   validConfig,
   deprecatedConfig,
-} = require('./fixtures/jest_config');
+} = require('./fixtures/jestConfig');
 
 test('recursively validates default Jest config', () => {
   expect(
@@ -210,7 +210,7 @@ test('works with custom deprecations', () => {
 
 test('works with multiple valid types', () => {
   const exampleConfig = {
-    foo: MultipleValidOptions('text', ['text']),
+    foo: multipleValidOptions('text', ['text']),
   };
 
   expect(
@@ -239,7 +239,7 @@ test('works with multiple valid types', () => {
 
 test('reports errors nicely when failing with multiple valid options', () => {
   const exampleConfig = {
-    foo: MultipleValidOptions('text', ['text']),
+    foo: multipleValidOptions('text', ['text']),
   };
 
   expect(() =>
@@ -254,7 +254,7 @@ test('reports errors nicely when failing with multiple valid options', () => {
 
 test('Repeated types within multiple valid examples are coalesced in error report', () => {
   const exampleConfig = {
-    foo: MultipleValidOptions('foo', 'bar', 2),
+    foo: multipleValidOptions('foo', 'bar', 2),
   };
 
   expect(() =>
@@ -265,4 +265,24 @@ test('Repeated types within multiple valid examples are coalesced in error repor
       },
     ),
   ).toThrowErrorMatchingSnapshot();
+});
+
+test('Comments in config JSON using "//" key are not warned', () => {
+  jest.spyOn(console, 'warn').mockImplementation(() => {});
+  const config = {'//': 'a comment'};
+
+  validate(config, {
+    exampleConfig: validConfig,
+  });
+  expect(console.warn).not.toBeCalled();
+
+  console.warn.mockReset();
+
+  validate(config, {
+    exampleConfig: validConfig,
+    recursiveBlacklist: [('myCustomKey': "don't validate this")],
+  });
+  expect(console.warn).not.toBeCalled();
+
+  console.warn.mockRestore();
 });